Support global patching
This adds support for patching the global fs object, which is
occasionally required at the application level when many other modules
throughout the system are set to use require('fs') instead of
require('graceful-fs').

Additionally, the fs builtin 'close()' and 'closeSync()' methods are
instrumented, so that a retry() is attempted on any close, rather than
only a close() by that particular instance of graceful-fs.