win: serialize read/write/appendFile and rename
rename can be used to atomically overwrite files in UNIX, but fails
in Windows when the destination is open. A file can be deleted while
open but will only be deleted on disk after it is closed, so a new
file cannot take its place right away. This change adds a queue,
making rename mutually exclusive with read/write/appendFile.