win32 rename EPERM: fail fast if target exists
When `rename` is used on Windows, and encounters an EPERM or EACCES
error, graceful-fs will retry for up to 30 seconds with gradual backoff
to get around file system locking due to A/V etc.

This was fixed in 59145cea32af980b675121ebc8c3255f780b7fc5

However, this is a problem if the target exists, which can cause an
EPERM or EACCES that is not due to folder locking.

This commit causes it to fail fast when an EPERM or EACCES is
raised by rename if the target exists.

Fix #98